chore: volume control components move
